Add swagger

+++ src/main/java/mabmab/retoibk/reto/infrastructure/controller/dto/PedidoRequest.java
@@ -5,6 +5,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -13,7 +14,6 @@ import java.time.LocalDate;
 public class PedidoRequest {
     @JsonFormat(pattern = "yyyy-MM-dd")
     private LocalDate fecha;
-    
-    private Long total;
+    private BigDecimal total;
     private boolean estado;
 }
